Updated to Yves 1.1.8 (see notes)
Yves' version still poll()s before asking the device to generateRemoteSensorEvents
This version ties takeAction() to changes in the devices runtimeRevision - when that changes, we know a poll() was just run/completed, so all we need to do is have the device generate the remote sensor events for us. ***MUCH*** less overhead, plus doesn't use a timer or schedule at all...clever, no?